Move LUP decomposition based circuit construction logic from `GF2MulK` to `SynthesizeLRCircuit` bloq (#1726)
The decomposition in `GF2MulK` works for any LR circuit and so it's
better to have it in the `SynthesizeLRCircuit` bloq. Both `GF2MulK` and
`GF2Square` will now have an explicit decomposition.

+++ qualtran/bloqs/gf_arithmetic/gf2_square.py
@@ -69,7 +69,7 @@ class GF2Square(Bloq):
         r"""$m \times m$ matrix that maps the input $x^{i}$ to $x^{2 * i} % P(x)$"""
         m = int(self.bitsize)
         f = self.qgf.gf_type.irreducible_poly
-        M = np.zeros((m, m))
+        M = np.zeros((m, m), dtype=int)
         alpha = [0] * m
         for i in range(m):
             # x ** (2 * i) % f
